third_party/WebKit/LayoutTests/fast/css/font-face-cache-version.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

async_test(function(t) {
  var messages = [];
  window.addEventListener('message', (evt) => {
    messages.push(evt.data);
    if (messages.length == 2) {
      t.step(() => {
        assert_not_equals(messages[0], messages[1],
                          'The texts in two iframes should not be identical');
        t.done();
      });
    }
  });
}, 'Cached font from different document should not be used');

</script>
<iframe id='frame' src='resources/font-face-cache-version-frame1.html'></iframe>
<iframe id='frame' src='resources/font-face-cache-version-frame2.html'></iframe>
